source file restructure
